requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
openai
gnews
yfinance
yahooquery
textblob
beautifulsoup4
duckduckgo-search
jinja2